+++ src/index.js
@@ -4,4 +4,4 @@ import App from './App'
 import reportWebVitals from './reportWebVitals'
 import 'semantic-ui-css/semantic.min.css'
 
-ReactDOM.render( < App / > , document.getElementById('root'))
+ReactDOM.render(<App />, document.getElementById('root'))
